Refactor: Cleaned a lot the api code and made it more secure

+++ src/app/api/collection/route.ts
@@ -1,30 +1,18 @@
-import { NextRequest, NextResponse } from "next/server";
+import { NextRequest } from "next/server";
 import prisma from "@/lib/prisma";
 import { auth } from "@/lib/auth";
+import { serverError, serverOk } from "@/lib/errorHandling/serverErrors";
 
 /** Creates a collection in the database */
 export const POST = async (req: NextRequest) => {
   try {
     const session = await auth();
-    if (session?.user?.id === undefined) {
-      return NextResponse.json(
-        { error: { message: "Not authenticated (no id found)" } },
-        { status: 401 }
-      );
-    }
+    if (session?.user?.id === undefined) return serverError("unauthenticated");
 
     const body = await req.json();
 
-    if (!body || typeof body !== "object" || !body.title) {
-      return NextResponse.json(
-        {
-          error: {
-            message: "Invalid payload. Title is required.",
-          },
-        },
-        { status: 400 }
-      );
-    }
+    if (!body || typeof body !== "object" || !body.title)
+      return serverError("invalid-payload", "Missing property: <title>");
 
     await prisma.collection.create({
       data: {
@@ -33,13 +21,9 @@ export const POST = async (req: NextRequest) => {
       },
     });
 
-    return NextResponse.json(null, { status: 200 });
+    return serverOk();
   } catch (err) {
-    console.error("Error while adding category :\n", err);
-    return NextResponse.json(
-      { error: { message: "failed-adding-to-db" } },
-      { status: 501 }
-    );
+    return serverError("internal-server-error", err);
   }
 };
 
@@ -49,36 +33,31 @@ export const PATCH = async (req: NextRequest) => {
     const action = params.get("action");
     const id = params.get("id");
 
-    if (!action || !id) {
-      return NextResponse.json(
-        { error: { message: "Properties action, id are needed" } },
-        { status: 400 }
-      );
-    }
+    if (!action)
+      return serverError("missing-parameters", "Parameter missing: <action>");
+    if (!id)
+      return serverError("missing-parameters", "Parameter missing: <id>");
 
     if (action === "rename") {
       const newtitle = params.get("newtitle");
 
-      if (!newtitle) {
-        return NextResponse.json(
-          { error: { message: "Property newtitle needed" } },
-          { status: 400 }
+      if (!newtitle)
+        return serverError(
+          "missing-parameters",
+          "Parameter missing: <newtitle>"
         );
-      }
 
       await prisma.collection.update({
         where: { id },
-        data: { title: newtitle },
+        data: { title: newtitle, updatedAt: new Date() },
       });
+      
+      return serverOk();
     }
 
-    return NextResponse.json(null, { status: 200 });
+    return serverError("invalid-patch-action");
   } catch (err) {
-    console.error("Error while deleting category :\n", err);
-    return NextResponse.json(
-      { error: { message: "failed-adding-to-db" } },
-      { status: 501 }
-    );
+    return serverError("internal-server-error", err);
   }
 };
 
@@ -86,22 +65,19 @@ export const DELETE = async (req: NextRequest) => {
   const params = req.nextUrl.searchParams;
   try {
     const id = params.get("id");
+    if (!id)
+      return serverError("missing-parameters", "Parameter missing: <id>");
 
-    if (!id) {
-      return NextResponse.json(
-        { error: { message: "missing-parameters" } },
-        { status: 400 }
-      );
-    }
+    const collection = await prisma.collection.findUnique({ where: { id } });
+
+    const session = await auth();
+    if (session?.user?.id !== collection?.ownerId)
+      return serverError("unauthorized");
 
     await prisma.collection.delete({ where: { id } });
 
-    return NextResponse.json(null, { status: 200 });
+    return serverOk();
   } catch (err) {
-    console.error("Error while deleting category :\n", err);
-    return NextResponse.json(
-      { error: { message: "failed-adding-to-db" } },
-      { status: 501 }
-    );
+    return serverError("internal-server-error", err);
   }
 };
